ASCII CAM Working
Got the ASCII CAM Working.  Works on mobile and desktop devices.
Discoverd that the width and height variabled in asciicam.js are vitally
imporant.They are set to 80 and 60 pixels respectively but each pixel is
represented by a character which actually taxes up more than one pixel,
hence the display to the user will be much more than that, about 6 times
actually.